requirements.txt
arrow==0.10.0
binaryornot==0.4.4
chardet==3.0.4
click==6.7
colander==1.4
cookiecutter==1.5.1
cornice==2.4.0
decorator==4.1.2
deform==2.0.4
future==0.16.0
hupper==1.0
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.10.2
Jinja2==2.9.6
jinja2-time==0.2.0
Mako==1.0.7
MarkupSafe==1.0

PasteDeploy==1.5.2
peppercorn==0.5
pexpect==4.2.1
pickleshare==0.7.4
plaster==0.5
plaster-pastedeploy==0.4.1
poyo==0.4.1
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
pyramid==1.9.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-tm==2.2
python-dateutil==2.6.1
repoze.lru==0.6
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
SQLAlchemy==1.1.13
traitlets==4.3.2
transaction==2.1.2
translationstring==1.3
typing==3.6.2
venusian==1.1.0
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.3
whichcraft==0.4.1
zope.deprecation==4.3.0
zope.interface==4.4.2
zope.sqlalchemy==0.7.7
requests==2.22.0
WebTest==2.0.33
pytest==5.3.4
